Remove SplashScreenActivity
