+++ main/AndroidManifest.xml
@@ -98,9 +98,13 @@
             android:value="3.0" />
 
         <activity
-            android:name=".SplashActivity"
+            android:name=".MainActivity"
             android:theme="@style/SplashScreenTheme"
-            android:exported="true" >
+            android:exported="true"
+            android:launchMode="singleTop"
+            android:windowSoftInputMode="adjustNothing">
+            <!-- 'adjustNothing' fixes the issue described in #11610 -->
+            <!-- don't add configChanges="orientation|screenSize" as we use different base layouts depending on the device orientation -->
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
@@ -112,15 +116,6 @@
             </intent-filter>
 
             <meta-data android:name="android.app.shortcuts" android:resource="@xml/shortcuts" />
-        </activity>
-        <activity
-            android:name=".MainActivity"
-            android:exported="true"
-            android:theme="@style/cgeo"
-            android:launchMode="singleTop"
-            android:windowSoftInputMode="adjustNothing" >
-            <!-- 'adjustNothing' fixes the issue described in #11610 -->
-            <!-- don't add configChanges="orientation|screenSize" as we use different base layouts depending on the device orientation -->
         </activity>
         <activity android:name=".DBInspectionActivity"
             android:exported="false"
